Excepted more things from auth token
Because the app is open to the public at this point, there's absolutely
no reason to require an authenticity token.  It makes it harder to use
the API, and doesn't actually buy you anything.  If we do end up adding
authentication, we'll need to do something fancy like OAuth for API
clients anyway, so for now it makes sense to just remove this.